+++ functions/fetch/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @enschedule-fns/fetch
 
+## 0.0.4
+
+### Patch Changes
+
+- Prefix env variables and fix bugs and make helm chart work with a migration job
+
 ## 0.0.3
 
 ### Patch Changes