+++ packages/worker-cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @enschedule/worker-cli
 
+## 0.0.4
+
+### Patch Changes
+
+- Prefix env variables and fix bugs and make helm chart work with a migration job
+- Updated dependencies
+  - @enschedule/worker@1.1.6
+
 ## 0.0.3
 
 ### Patch Changes